fixes to sample code:
ftdi: gets a list of FTDI devices and lets the user choose which one to open, reads multiple lines to allow *ok and continuous mode to not mess up the buffer.
i2c: adds a new command which creates a list of i2c addresses on the line
uart: reads multiple lines to allow *ok and continuous mode to not mess up the buffer.
readme: updated documentation to work with these new features